+++ app/controllers/PsaSchemeTaskListController.scala
@@ -15,8 +15,8 @@
  */
 
 package controllers
-
 import config.FrontendAppConfig
+import connectors.SchemeDetailsConnector
 import connectors.MinimalPsaConnector
 import controllers.actions._
 import identifiers.SchemeNameId
@@ -34,59 +34,66 @@ import views.html.psaTaskList
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
+import utils.annotations.Racdac
+
+
+import scala.concurrent.{ExecutionContext, Future}
 
 class PsaSchemeTaskListController @Inject()(appConfig: FrontendAppConfig,
+                                            schemeDetailsConnector: SchemeDetailsConnector,
                                          override val messagesApi: MessagesApi,
                                          authenticate: AuthAction,
-                                         getData: DataRetrievalAction,
-                                         minimalPsaConnector: MinimalPsaConnector,
-                                         @TaskList allowAccess: AllowAccessActionProvider,
-                                         val controllerComponents: MessagesControllerComponents,
-                                         val view: psaTaskList,
-                                         hsTaskListHelperRegistration: HsTaskListHelperRegistration,
-                                         hsTaskListHelperVariations: HsTaskListHelperVariations
+                                         val controllerComponents: MessagesControllerComponents
                                         )(implicit val executionContext: ExecutionContext) extends
   FrontendBaseController with I18nSupport with Retrievals {
 
-  private def redirects(implicit request: OptionalDataRequest[AnyContent], hc: HeaderCarrier):Future[Option[Result]] = {
-    request.psaId match {
-      case None => Future.successful(None)
-      case Some(psaId) =>
-        minimalPsaConnector.getMinimalFlags(psaId.id).map {
-          case PSAMinimalFlags(_, true, false) => Some(Redirect(Call("GET", appConfig.youMustContactHMRCUrl)))
-          case PSAMinimalFlags(_, false, true) => Some(Redirect(Call("GET",appConfig.psaUpdateContactDetailsUrl)))
-          case _ => None
+
+
+
+  def onPageLoad(mode: Mode, srn: Option[String]): Action[AnyContent] = authenticate(Some(PSA)).async { implicit request =>
+    (request.psaId, srn) match {
+      case (Some(psaId), Some(srnNo)) =>
+        schemeDetailsConnector.getSchemeDetails(psaId.id, "srn", srnNo).map {ua =>
+          ua.get(IsRacDacId) match {
+            case Some(true) =>
+              Redirect(controllers.routes.PsaRacdacSchemeTaskListController.onPageLoad(mode,srn))
+            case Some(false) =>
+              Redirect(controllers.routes.PsaNormalSchemeTaskListController.onPageLoad(mode,srn))
+            case _ =>
+              Redirect(controllers.routes.SessionExpiredController.onPageLoad())
+          }
         }
+      case _ =>
+        Future(Redirect(controllers.routes.SessionExpiredController.onPageLoad()))
     }
   }
 
-  def onPageLoad(mode: Mode, srn: Option[String]): Action[AnyContent] = (authenticate(Some(PSA)) andThen getData(mode, srn, refreshData = true)
-    andThen allowAccess(srn)).async {
-    implicit request =>
-
-      redirects.map {
-        case Some(result) => result
-        case _ =>
-          val schemeNameOpt: Option[String] = request.userAnswers.flatMap(_.get(SchemeNameId))
-          val isRacDacOpt: Option[Boolean] = request.userAnswers.flatMap(_.get(IsRacDacId))
-
-          (srn, request.userAnswers, schemeNameOpt, isRacDacOpt) match {
-
-            case (_, Some(_), Some(_), Some(true)) =>
-              Redirect(controllers.racdac.routes.CheckYourAnswersController.onPageLoad(UpdateMode, srn))
-
-            case (None, Some(userAnswers), Some(schemeName), _) =>
-              Ok(view(hsTaskListHelperRegistration.taskList(userAnswers, None, srn), schemeName))
-
-            case (Some(_), Some(userAnswers), Some(schemeName), _) =>
-              Ok(view(hsTaskListHelperVariations.taskList(userAnswers, Some(request.viewOnly), srn), schemeName))
-
-            case (Some(_), _, _, _) =>
-              Redirect(controllers.routes.SessionExpiredController.onPageLoad())
-
-            case _ =>
-              Redirect(appConfig.managePensionsSchemeOverviewUrl)
-          }
-      }
-  }
+//  def onPageLoad(mode: Mode, srn: Option[String]): Action[AnyContent] = authenticate(Some(PSA)).async {
+//    implicit request =>
+//
+//      redirects.map {
+//        case Some(result) => result
+//        case _ =>
+//          val schemeNameOpt: Option[String] = request.userAnswers.flatMap(_.get(SchemeNameId))
+//          val isRacDacOpt: Option[Boolean] = request.userAnswers.flatMap(_.get(IsRacDacId))
+//
+//          (srn, request.userAnswers, schemeNameOpt, isRacDacOpt) match {
+//
+//            case (_, Some(_), Some(_), Some(true)) =>
+//              Redirect(controllers.racdac.routes.CheckYourAnswersController.onPageLoad(UpdateMode, srn))
+//
+//            case (None, Some(userAnswers), Some(schemeName), _) =>
+//              Ok(view(hsTaskListHelperRegistration.taskList(userAnswers, None, srn), schemeName))
+//
+//            case (Some(_), Some(userAnswers), Some(schemeName), _) =>
+//              Ok(view(hsTaskListHelperVariations.taskList(userAnswers, Some(request.viewOnly), srn), schemeName))
+//
+//            case (Some(_), _, _, _) =>
+//              Redirect(controllers.routes.SessionExpiredController.onPageLoad())
+//
+//            case _ =>
+//              Redirect(appConfig.managePensionsSchemeOverviewUrl)
+//          }
+//      }
+//  }
 }
